Refactor: 점진적 마이그레이션(스피너)
